build/rules.mk: Fix the `all` target running twice
If a rule has no body, make considers it a dependency-only rule, and will run
any other rules that match our target, after running the dependency. In
practice, this meant that `make all` - which appears to be the default - ran
both `kaleidoscope-builder build-all`, and `kaleidoscope-builder all` (due to
the catch-all `%` rule at the end). The latter is an error.

This little change adds a dummy body to the `all` rule, so that the `%` will not
be considered by make.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>