Fix the library ordering when building outside of Arduino-Boards
When building outside of Arduino-Boards, we want to use the current library, not
whatever version is in Arduino-Boards at the time. To do so, we must tell
`arduino-builder` about the library, because the current directory is not
considered by default.

Now, we can't use `-libraries .`, because `arduino-builder` will search for a
directory named like the library, and we don't have that. We are already in the
library directory at this point.

So, we need a library outside of the current dir. We could use `..`, but we want
to be safe, to only have the current library there. For this reason, the
`travis-smoke-examples` target will create a `current-libraries` directory one
level up, symlink the current directory there, under the library name, and add
`-libraries $(pwd)/../current-libraries` to the `arduino-builder` arguments.

All of this makes the builder consider the current library the best to use.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>